Add Bouncing Anim to make a proper request on GB
This is from a unused animation in Sonic The Hedgehog (1991). A few weeks ago a SHC mod called Sonic is alone appeared, and it makes sonic actually have the ability to "bounce" when he falls down the map or when you press the A button. This is my best recreation with it's palette and everything.